pyproject.toml
[project]
name = "unlabeled"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.0.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "tensorboard>=2.20.0",
    "torch==2.9.1",
    "torchvision==0.24.1",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
torch = [
    { index = "torch-cu129" },
]
torchvison = [
    { index = "torch-cu129" },
]

[[tool.uv.index]]
name = "torch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[dependency-groups]
dev = [
    "mypy>=1.18.2",
]
